refactor: Move test files alongside their source files
Improve test file organization by co-locating tests with source files:

Moved files:
- StatefulChart.test.tsx → src/chart/components/ (alongside StatefulChart.tsx)
- MatrixifyGridRenderer.test.tsx → src/chart/components/Matrixify/ (alongside MatrixifyGridRenderer.tsx)
- matrixify.test.ts → src/chart/types/ (alongside matrixify.ts)
- __mocks__/matrixify.ts → src/chart/types/matrixify.mocks.ts (renamed and co-located)

Updated import paths in moved test files to reflect new locations.
Removed empty __mocks__ directory.

Benefits:
- Better test discoverability - tests next to source files
- Easier maintenance - related files grouped together
- Cleaner test directory structure
- Follows common testing patterns

All tests continue to pass with updated paths.
