+++ superset-frontend/packages/superset-ui-core/src/chart/components/Matrixify/MatrixifyGridRenderer.test.tsx
@@ -17,336 +17,380 @@
  * under the License.
  */
 
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { ThemeProvider, supersetTheme } from '../../..';
+import { ThemeProvider } from '@superset-ui/core';
 import MatrixifyGridRenderer from './MatrixifyGridRenderer';
-import { MatrixifyGrid } from '../../types/matrixify';
-import { AdhocMetric } from '../../../query/types/Metric';
-
 import { generateMatrixifyGrid } from './MatrixifyGridGenerator';
+import { supersetTheme } from '../../../theme';
 
-// Use 'any' to bypass strict typing in tests
-type TestFormData = any;
-
-// Mock MatrixifyGridCell component
-jest.mock('./MatrixifyGridCell', () => ({
-  __esModule: true,
-  default: ({ cell }: any) => (
-    <div data-testid={`cell-${cell.id}`} className="matrixify-cell">
-      {cell.title || `${cell.rowLabel}-${cell.colLabel}`}
-    </div>
-  ),
-}));
-
-// Mock generateMatrixifyGrid function
+// Mock the MatrixifyGridGenerator
 jest.mock('./MatrixifyGridGenerator', () => ({
   generateMatrixifyGrid: jest.fn(),
 }));
 
-const mockedGenerateMatrixifyGrid =
-  generateMatrixifyGrid as jest.MockedFunction<typeof generateMatrixifyGrid>;
+// Mock MatrixifyGridCell component
+jest.mock('./MatrixifyGridCell', () =>
+  // eslint-disable-next-line react/display-name, @typescript-eslint/no-unused-vars
+  ({ cell, rowHeight, datasource, hooks }: any) => (
+    <div data-testid={`grid-cell-${cell.id}`}>Cell: {cell.id}</div>
+  ),
+);
 
-const mockDatasource = {
-  id: 1,
-  type: 'table' as const,
-  uid: '1__table',
-  datasource_name: 'test_datasource',
-  table_name: 'test_table',
-  database: {
-    id: 1,
-    name: 'test_database',
-  },
-};
-
-const createAdhocMetric = (label: string): AdhocMetric => ({
-  expressionType: 'SIMPLE',
-  column: { column_name: 'value' },
-  aggregate: 'SUM',
-  label,
-});
-
-const baseFormData: TestFormData = {
-  viz_type: 'table',
-  datasource: '1__table',
-  matrixify_mode_rows: 'metrics',
-  matrixify_mode_columns: 'metrics',
-  matrixify_rows: [createAdhocMetric('Revenue'), createAdhocMetric('Profit')],
-  matrixify_columns: [createAdhocMetric('Q1'), createAdhocMetric('Q2')],
-  matrixify_show_row_labels: true,
-  matrixify_show_column_headers: true,
-  matrixify_row_height: 200,
-};
-
-const mockGrid: MatrixifyGrid = {
-  rowHeaders: ['Revenue', 'Profit'],
-  colHeaders: ['Q1', 'Q2'],
-  cells: [
-    [
-      {
-        id: 'matrixify-0-0',
-        row: 0,
-        col: 0,
-        rowLabel: 'Revenue',
-        colLabel: 'Q1',
-        title: 'Revenue - Q1',
-        formData: { viz_type: 'big_number_total' },
-      },
-      {
-        id: 'matrixify-0-1',
-        row: 0,
-        col: 1,
-        rowLabel: 'Revenue',
-        colLabel: 'Q2',
-        title: 'Revenue - Q2',
-        formData: { viz_type: 'big_number_total' },
-      },
-    ],
-    [
-      {
-        id: 'matrixify-1-0',
-        row: 1,
-        col: 0,
-        rowLabel: 'Profit',
-        colLabel: 'Q1',
-        title: 'Profit - Q1',
-        formData: { viz_type: 'big_number_total' },
-      },
-      {
-        id: 'matrixify-1-1',
-        row: 1,
-        col: 1,
-        rowLabel: 'Profit',
-        colLabel: 'Q2',
-        title: 'Profit - Q2',
-        formData: { viz_type: 'big_number_total' },
-      },
-    ],
-  ],
-  baseFormData,
-};
-
-const defaultProps = {
-  formData: baseFormData,
-  width: 800,
-  height: 600,
-  datasource: mockDatasource,
-};
+const mockGenerateMatrixifyGrid = generateMatrixifyGrid as jest.MockedFunction<
+  typeof generateMatrixifyGrid
+>;
 
 const renderWithTheme = (component: React.ReactElement) =>
   render(<ThemeProvider theme={supersetTheme}>{component}</ThemeProvider>);
 
 beforeEach(() => {
-  mockedGenerateMatrixifyGrid.mockReturnValue(mockGrid);
-});
-
-afterEach(() => {
   jest.clearAllMocks();
 });
 
-test('should render the grid with all cells', () => {
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  expect(screen.getByText('Revenue - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Revenue - Q2')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q2')).toBeInTheDocument();
-});
-
-test('should render column headers when enabled', () => {
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  expect(screen.getByText('Q1')).toBeInTheDocument();
-  expect(screen.getByText('Q2')).toBeInTheDocument();
-});
-
-test('should render row labels when enabled', () => {
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  const rowLabels = screen.getAllByText('Revenue');
-  expect(rowLabels).toHaveLength(1); // Only in row label, not in cells
-
-  const profitLabels = screen.getAllByText('Profit');
-  expect(profitLabels).toHaveLength(1);
-});
-
-test('should hide column headers when disabled', () => {
-  const formDataWithoutHeaders = {
-    ...baseFormData,
-    matrixify_show_column_headers: false,
-  };
-
-  renderWithTheme(
-    <MatrixifyGridRenderer
-      {...defaultProps}
-      formData={formDataWithoutHeaders}
-    />,
-  );
-
-  expect(screen.queryByText('Q1')).not.toBeInTheDocument();
-  expect(screen.queryByText('Q2')).not.toBeInTheDocument();
-});
-
-test('should hide row labels when disabled', () => {
-  const formDataWithoutLabels = {
-    ...baseFormData,
-    matrixify_show_row_labels: false,
-  };
-
-  renderWithTheme(
-    <MatrixifyGridRenderer
-      {...defaultProps}
-      formData={formDataWithoutLabels}
-    />,
-  );
-
-  // Row labels should not be shown
-  expect(screen.queryAllByText('Revenue')).toHaveLength(0);
-  expect(screen.queryAllByText('Profit')).toHaveLength(0);
-});
-
-test('should calculate column widths based on grid layout', () => {
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  // Verify all cells are rendered
-  expect(screen.getByText('Revenue - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Revenue - Q2')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q2')).toBeInTheDocument();
-});
-
-test('should handle custom row height', () => {
-  const formDataWithCustomHeight = {
-    ...baseFormData,
-    matrixify_row_height: 300,
-  };
-
-  renderWithTheme(
-    <MatrixifyGridRenderer
-      {...defaultProps}
-      formData={formDataWithCustomHeight}
-    />,
-  );
-
-  // Verify cells are rendered (height would be applied in the actual component)
-  expect(screen.getByText('Revenue - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Revenue - Q2')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q2')).toBeInTheDocument();
-});
-
-test('should handle empty grid', () => {
-  mockedGenerateMatrixifyGrid.mockReturnValue({
-    rowHeaders: [],
-    colHeaders: [],
-    cells: [],
-    baseFormData,
-  });
-
-  const { container } = renderWithTheme(
-    <MatrixifyGridRenderer {...defaultProps} />,
-  );
-
-  // Should render empty container
-  expect(container.firstChild).toBeInTheDocument();
-  expect(screen.queryAllByText(/Revenue|Profit|Q1|Q2/)).toHaveLength(0);
-});
-
-test('should handle single cell grid', () => {
-  mockedGenerateMatrixifyGrid.mockReturnValue({
-    rowHeaders: ['Total'],
-    colHeaders: ['All'],
+test('should create single group when fitting columns dynamically', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1', 'Row 2'],
+    colHeaders: ['Col 1', 'Col 2', 'Col 3', 'Col 4', 'Col 5'],
     cells: [
       [
-        {
-          id: 'matrixify-0-0',
-          row: 0,
-          col: 0,
-          rowLabel: 'Total',
-          colLabel: 'All',
-          formData: { viz_type: 'big_number_total' },
-        },
+        { id: 'cell-0-0' },
+        { id: 'cell-0-1' },
+        { id: 'cell-0-2' },
+        { id: 'cell-0-3' },
+        { id: 'cell-0-4' },
+      ],
+      [
+        { id: 'cell-1-0' },
+        { id: 'cell-1-1' },
+        { id: 'cell-1-2' },
+        { id: 'cell-1-3' },
+        { id: 'cell-1-4' },
       ],
     ],
-    baseFormData,
-  });
+  };
 
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
 
-  expect(screen.getByText('Total-All')).toBeInTheDocument();
-  expect(screen.getByText('Total')).toBeInTheDocument();
-  expect(screen.getByText('All')).toBeInTheDocument();
-});
-
-test('should handle grid with null cells', () => {
-  mockedGenerateMatrixifyGrid.mockReturnValue({
-    rowHeaders: ['A', 'B'],
-    colHeaders: ['X', 'Y'],
-    cells: [
-      [mockGrid.cells[0][0], null],
-      [null, mockGrid.cells[1][1]],
-    ],
-    baseFormData,
-  });
-
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  // Only non-null cells should be rendered
-  expect(screen.getByText('Revenue - Q1')).toBeInTheDocument();
-  expect(screen.getByText('Profit - Q2')).toBeInTheDocument();
-  expect(screen.queryByText('null')).not.toBeInTheDocument();
-});
-
-test('should apply overflow styling to container', () => {
-  const { container } = renderWithTheme(
-    <MatrixifyGridRenderer {...defaultProps} />,
-  );
-
-  const gridContainer = container.firstChild as HTMLElement;
-  expect(gridContainer).toHaveStyle({
-    width: '100%',
-    height: '600px',
-  });
-});
-
-test('should call generateMatrixifyGrid with correct formData', () => {
-  renderWithTheme(<MatrixifyGridRenderer {...defaultProps} />);
-
-  expect(mockedGenerateMatrixifyGrid).toHaveBeenCalledTimes(1);
-  expect(mockedGenerateMatrixifyGrid).toHaveBeenCalledWith(baseFormData);
-});
-
-test('should handle dimension mode grid', () => {
-  const dimensionFormData: TestFormData = {
-    viz_type: 'table',
-    datasource: '1__table',
-    matrixify_mode_rows: 'dimensions',
-    matrixify_mode_columns: 'dimensions',
-    matrixify_dimension_rows: {
-      dimension: 'country',
-      values: ['USA', 'Canada'],
-    },
-    matrixify_dimension_columns: {
-      dimension: 'product',
-      values: ['Widget', 'Gadget'],
-    },
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: true,
+    matrixify_charts_per_row: 3,
     matrixify_show_row_labels: true,
     matrixify_show_column_headers: true,
   };
 
-  mockedGenerateMatrixifyGrid.mockReturnValue({
-    rowHeaders: ['USA', 'Canada'],
-    colHeaders: ['Widget', 'Gadget'],
-    cells: mockGrid.cells,
-    baseFormData: dimensionFormData,
-  });
-
-  renderWithTheme(
-    <MatrixifyGridRenderer {...defaultProps} formData={dimensionFormData} />,
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
   );
 
-  expect(screen.getByText('USA')).toBeInTheDocument();
-  expect(screen.getByText('Canada')).toBeInTheDocument();
-  expect(screen.getByText('Widget')).toBeInTheDocument();
-  expect(screen.getByText('Gadget')).toBeInTheDocument();
+  // When fitting dynamically, should have only one column group with all 5 columns
+  // Check for the presence of the grid structure
+  const gridContainers = container.querySelectorAll('div[class*="css-"]');
+  expect(gridContainers.length).toBeGreaterThan(0);
+
+  // Verify all 5 column headers are present in single group
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders).toHaveLength(5);
+  expect(columnHeaders[0]).toHaveTextContent('Col 1');
+  expect(columnHeaders[4]).toHaveTextContent('Col 5');
+});
+
+test('should create multiple groups when not fitting columns dynamically', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1', 'Row 2'],
+    colHeaders: ['Col 1', 'Col 2', 'Col 3', 'Col 4', 'Col 5'],
+    cells: [
+      [
+        { id: 'cell-0-0' },
+        { id: 'cell-0-1' },
+        { id: 'cell-0-2' },
+        { id: 'cell-0-3' },
+        { id: 'cell-0-4' },
+      ],
+      [
+        { id: 'cell-1-0' },
+        { id: 'cell-1-1' },
+        { id: 'cell-1-2' },
+        { id: 'cell-1-3' },
+        { id: 'cell-1-4' },
+      ],
+    ],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: false,
+    matrixify_charts_per_row: 3,
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // With 5 columns and charts_per_row=3, should have 2 groups (3+2)
+  // With 2 rows and wrapping, we should see headers repeated
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders.length).toBeGreaterThanOrEqual(5); // At least the base headers
+});
+
+test('should handle exact division of columns', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1'],
+    colHeaders: ['Col 1', 'Col 2', 'Col 3', 'Col 4'],
+    cells: [
+      [
+        { id: 'cell-0-0' },
+        { id: 'cell-0-1' },
+        { id: 'cell-0-2' },
+        { id: 'cell-0-3' },
+      ],
+    ],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: false,
+    matrixify_charts_per_row: 2,
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // With 4 columns and charts_per_row=2, should have exactly 2 groups (2+2)
+  // Check that we have column headers - should be 4 total (2 per group)
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders).toHaveLength(4);
+});
+
+test('should handle case where charts_per_row exceeds total columns', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1'],
+    colHeaders: ['Col 1', 'Col 2'],
+    cells: [[{ id: 'cell-0-0' }, { id: 'cell-0-1' }]],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: false,
+    matrixify_charts_per_row: 5,
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // Should create only one group with all columns
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders).toHaveLength(2);
+});
+
+test('should show headers for each group when wrapping occurs', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1', 'Row 2'],
+    colHeaders: ['Col 1', 'Col 2', 'Col 3'],
+    cells: [
+      [{ id: 'cell-0-0' }, { id: 'cell-0-1' }, { id: 'cell-0-2' }],
+      [{ id: 'cell-1-0' }, { id: 'cell-1-1' }, { id: 'cell-1-2' }],
+    ],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: false,
+    matrixify_charts_per_row: 2,
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // With wrapping (multiple column groups), headers should appear for each group
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders.length).toBeGreaterThan(3); // More than just one set of headers
+
+  // Row headers should appear only once per row (for first column group)
+  const rowHeaders = container.querySelectorAll('.matrixify-row-header');
+  expect(rowHeaders).toHaveLength(2); // One for each row
+});
+
+test('should show headers only on first row when not wrapping', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1', 'Row 2'],
+    colHeaders: ['Col 1', 'Col 2'],
+    cells: [
+      [{ id: 'cell-0-0' }, { id: 'cell-0-1' }],
+      [{ id: 'cell-1-0' }, { id: 'cell-1-1' }],
+    ],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: true, // No wrapping
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // Without wrapping, headers should appear only once (first row)
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders).toHaveLength(2); // Just Col 1 and Col 2
+
+  const rowHeaders = container.querySelectorAll('.matrixify-row-header');
+  expect(rowHeaders).toHaveLength(2); // One for each row
+});
+
+test('should hide headers when disabled', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1'],
+    colHeaders: ['Col 1', 'Col 2'],
+    cells: [[{ id: 'cell-0-0' }, { id: 'cell-0-1' }]],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_show_row_labels: false,
+    matrixify_show_column_headers: false,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  const columnHeaders = container.querySelectorAll('.matrixify-col-header');
+  expect(columnHeaders).toHaveLength(0);
+
+  const rowHeaders = container.querySelectorAll('.matrixify-row-header');
+  expect(rowHeaders).toHaveLength(0);
+});
+
+test('should place cells correctly in wrapped layout', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1'],
+    colHeaders: ['Col 1', 'Col 2', 'Col 3'],
+    cells: [[{ id: 'cell-0-0' }, { id: 'cell-0-1' }, { id: 'cell-0-2' }]],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    matrixify_fit_columns_dynamically: false,
+    matrixify_charts_per_row: 2,
+    matrixify_show_row_labels: true,
+    matrixify_show_column_headers: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // All cells should be rendered
+  const cells = container.querySelectorAll('[data-testid^="grid-cell-"]');
+  expect(cells).toHaveLength(3);
+  expect(
+    container.querySelector('[data-testid="grid-cell-cell-0-0"]'),
+  ).toBeInTheDocument();
+  expect(
+    container.querySelector('[data-testid="grid-cell-cell-0-1"]'),
+  ).toBeInTheDocument();
+  expect(
+    container.querySelector('[data-testid="grid-cell-cell-0-2"]'),
+  ).toBeInTheDocument();
+});
+
+test('should handle null grid gracefully', () => {
+  mockGenerateMatrixifyGrid.mockReturnValue(null);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  expect(container).toBeEmptyDOMElement();
+});
+
+test('should handle empty grid gracefully', () => {
+  const mockGrid: any = {
+    rowHeaders: [],
+    colHeaders: [],
+    cells: [],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // Should render container but no cells
+  expect(container).not.toBeEmptyDOMElement();
+  const gridCells = container.querySelectorAll('[data-testid^="grid-cell-"]');
+  expect(gridCells).toHaveLength(0);
+});
+
+test('should use default values for missing configuration', () => {
+  const mockGrid: any = {
+    rowHeaders: ['Row 1'],
+    colHeaders: ['Col 1', 'Col 2'],
+    cells: [[{ id: 'cell-0-0' }, { id: 'cell-0-1' }]],
+  };
+
+  mockGenerateMatrixifyGrid.mockReturnValue(mockGrid);
+
+  const formData = {
+    viz_type: 'test_chart',
+    matrixify_enabled: true,
+    // Missing optional configurations
+  };
+
+  const { container } = renderWithTheme(
+    <MatrixifyGridRenderer formData={formData} />,
+  );
+
+  // Should still render with defaults
+  expect(container).not.toBeEmptyDOMElement();
+  const gridCells = container.querySelectorAll('[data-testid^="grid-cell-"]');
+  expect(gridCells).toHaveLength(2);
 });
